# shared/openapi/map-service.yaml
openapi: 3.0.3
info:
  title: API микросервиса карт Taxi MVP
  version: 1.0.0
  description: "Микросервис карт для выдачи тайлов, локального кеширования PNG-тайлов, startup-загрузки OSM PBF Туркменистана, ночной синхронизации карты, manifest и delta-обновлений."
servers:
  - url: http://localhost:8090
paths:
...
  /api/map/manifest:
    get:
      summary: Manifest тайлов
      operationId: getMapManifest
      description: "Возвращает manifest тайлов для региона: поддерживаемые zoom levels, checksum значения, URL template и час ночной синхронизации карты."
      parameters:
        - name: region
          in: query
          description: "Регион карты. Если параметр не передан, используется turkmenistan."
          required: false
          schema:
            type: string
            default: turkmenistan
      responses:
        "200":
          description: Группы тайлов и checksum для синхронизации кеша клиента.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapManifestResponse"
  /api/map/delta:
    get:
      summary: Delta-обновление карты
      operationId: getMapDelta
      description: "Возвращает список измененных и удаленных тайлов между двумя версиями карты. Нужен, чтобы мобильное приложение скачивало только изменения, а не всю карту заново."
      parameters:
        - name: from
          in: query
          description: Версия карты, которая уже сохранена на устройстве.
          required: false
          schema:
            type: string
          example: tm-2026.05-demo
        - name: to
          in: query
          description: Целевая версия карты. Если не передана, используется текущая версия сервиса.
          required: false
          schema:
            type: string
          example: tm-2026.06-demo
      responses:
        "200":
          description: Измененные и удаленные тайлы между версиями.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapDeltaResponse"
...
  /tiles/{z}/{x}/{y}.png:
    get:
      summary: PNG-тайл карты
      operationId: getRasterTile
      description: "Возвращает PNG raster tile по координатам web map tile scheme. Сервис сначала ищет тайл в локальном `/data/tiles`; если тайла нет, скачивает его из OpenStreetMap, сохраняет локально и отдает клиенту."
      parameters:
        - name: z
          in: path
          description: Zoom level карты.
          required: true
          schema:
            type: integer
          example: 10
        - name: x
          in: path
          description: Номер колонки tile.
          required: true
          schema:
            type: integer
          example: 637
        - name: y
          in: path
          description: Номер строки tile.
          required: true
          schema:
            type: integer
          example: 412
      responses:
        "200":
          description: PNG-тайл карты.
          content:
            image/png:
              schema:
                type: string
                format: binary
        "502":
          description: Внешний tile provider недоступен.
components:
  schemas:
...
    MapManifestResponse:
      type: object
      description: Manifest для кеширования карты на клиенте.
      properties:
        region:
          type: string
          description: Регион карты.
          example: turkmenistan
        version:
          type: string
          description: Версия карты.
          example: tm-2026.06-demo
        strategy:
          type: string
          description: Стратегия обновления карты.
          example: delta
        nightlySyncHour:
          type: integer
          description: Рекомендуемый час фоновой синхронизации.
          example: 3
        tiles:
          type: array
          description: Группы тайлов для загрузки и проверки кеша.
          items:
            $ref: "#/components/schemas/TileGroup"
    TileGroup:
      type: object
      description: Группа тайлов одного zoom level.
      properties:
        z:
          type: integer
          description: Zoom level.
          example: 10
        checksum:
          type: string
          description: Контрольная сумма группы тайлов.
        urlTemplate:
          type: string
          description: URL-шаблон для загрузки PNG-тайлов.
          example: /tiles/{z}/{x}/{y}.png
    MapDeltaResponse:
      type: object
      description: Delta-обновление между версиями карты.
      properties:
        from:
          type: string
          description: Старая версия карты.
        to:
          type: string
          description: Новая версия карты.
        changed:
          type: array
          description: Тайлы, которые нужно скачать заново.
          items:
            $ref: "#/components/schemas/ChangedTile"
        deleted:
          type: array
          description: Тайлы, которые нужно удалить из локального кеша.
          items:
            type: string
    ChangedTile:
      type: object
      description: Измененный tile.
      properties:
        z:
          type: integer
          description: Zoom level.
        x:
          type: integer
          description: Колонка tile.
        y:
          type: integer
          description: Строка tile.
        checksum:
          type: string
          description: Новый checksum tile.
...